src/ml_pipeline/requirements.txt
# ML Pipeline Requirements

# Core dependencies
pyyaml>=6.0

# ML frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Data handling
pandas>=1.5.0
scikit-learn>=1.2.0

# Utilities
tqdm>=4.65.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
